+++ catalog/files/api/src/main/java/org/projectnessie/catalog/files/config/GcsNamedBucketOptions.java
@@ -18,26 +18,13 @@ package org.projectnessie.catalog.files.config;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import java.util.Optional;
 import org.immutables.value.Value;
 
 @Value.Immutable
 @JsonSerialize(as = ImmutableGcsNamedBucketOptions.class)
 @JsonDeserialize(as = ImmutableGcsNamedBucketOptions.class)
 @SuppressWarnings("immutables:subtype")
-public interface GcsNamedBucketOptions extends GcsBucketOptions {
-
-  GcsBucketOptions FALLBACK = ImmutableGcsNamedBucketOptions.builder().build();
-
-  /**
-   * The name of the bucket. If unset, the name of the bucket will be extracted from the
-   * configuration option, e.g. if {@code nessie.catalog.service.gcs.bucket1.name=my-bucket} is set,
-   * the bucket name will be {@code my-bucket}; otherwise, it will be {@code bucket1}.
-   *
-   * <p>This should only be defined if the bucket name contains non-alphanumeric characters, such as
-   * dots or dashes.
-   */
-  Optional<String> name();
+public interface GcsNamedBucketOptions extends GcsBucketOptions, PerBucket {
 
   @Value.NonAttribute
   @JsonIgnore
@@ -45,6 +32,8 @@ public interface GcsNamedBucketOptions extends GcsBucketOptions {
     return ImmutableGcsNamedBucketOptions.builder()
         .from(GcsBucketOptions.super.deepClone())
         .name(name())
+        .authority(authority())
+        .pathPrefix(pathPrefix())
         .build();
   }
 }