+++ catalog/files/api/src/main/java/org/projectnessie/catalog/files/config/GcsOptions.java
@@ -15,16 +15,18 @@
  */
 package org.projectnessie.catalog.files.config;
 
+import static org.projectnessie.catalog.files.config.OptionsUtil.resolveSpecializedBucket;
+
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
 import com.fasterxml.jackson.databind.annotation.JsonSerialize;
-import java.util.HashMap;
 import java.util.Map;
 import java.util.Optional;
 import org.immutables.value.Value;
 import org.projectnessie.nessie.docgen.annotations.ConfigDocs.ConfigItem;
 import org.projectnessie.nessie.docgen.annotations.ConfigDocs.ConfigPropertyName;
 import org.projectnessie.nessie.immutables.NessieImmutable;
+import org.projectnessie.storage.uri.StorageUri;
 
 /**
  * Configuration for Google Cloud Storage (GCS) object stores.
@@ -48,56 +50,25 @@ public interface GcsOptions {
 
   /**
    * Per-bucket configurations. The effective value for a bucket is taken from the per-bucket
-   * setting. If no per-bucket setting is present, uses the defaults from the top-level GCS
-   * settings.
+   * setting. If no per-bucket setting is present, uses the defaults from the top-level GCS settings
+   * in {@code default-options}.
    */
   @ConfigItem(section = "buckets")
-  @ConfigPropertyName("bucket-name")
+  @ConfigPropertyName("key")
   Map<String, GcsNamedBucketOptions> buckets();
 
-  default GcsBucketOptions effectiveOptionsForBucket(Optional<String> bucketName) {
-    GcsBucketOptions defaultOptions = defaultOptions().orElse(GcsNamedBucketOptions.FALLBACK);
+  default GcsNamedBucketOptions resolveOptionsForUri(StorageUri uri) {
+    Optional<GcsNamedBucketOptions> specific = resolveSpecializedBucket(uri, buckets());
 
-    if (bucketName.isEmpty()) {
-      return defaultOptions;
-    }
-
-    GcsBucketOptions specific = buckets().get(bucketName.get());
-    if (specific == null) {
-      return defaultOptions;
-    }
-
-    return ImmutableGcsNamedBucketOptions.builder().from(defaultOptions).from(specific).build();
+    ImmutableGcsNamedBucketOptions.Builder builder = ImmutableGcsNamedBucketOptions.builder();
+    defaultOptions().ifPresent(builder::from);
+    specific.ifPresent(builder::from);
+    return builder.build();
   }
 
   default GcsOptions validate() {
     // nothing to validate
     return this;
-  }
-
-  @Value.Check
-  default GcsOptions normalizeBuckets() {
-    Map<String, GcsNamedBucketOptions> buckets = new HashMap<>();
-    for (String bucketName : buckets().keySet()) {
-      GcsNamedBucketOptions options = buckets().get(bucketName);
-      if (options.name().isPresent()) {
-        bucketName = options.name().get();
-      } else {
-        options = ImmutableGcsNamedBucketOptions.builder().from(options).name(bucketName).build();
-      }
-      if (buckets.put(bucketName, options) != null) {
-        throw new IllegalArgumentException(
-            "Duplicate GCS bucket name '" + bucketName + "', check your GCS bucket configurations");
-      }
-    }
-    if (buckets.equals(buckets())) {
-      return this;
-    }
-    return ImmutableGcsOptions.builder()
-        .from(this)
-        .defaultOptions(defaultOptions())
-        .buckets(buckets)
-        .build();
   }
 
   @Value.NonAttribute